:building_construction: migrate all hooks and components to TS format
